+++ modules/apps/roles-admin/roles-admin-web/src/com/liferay/roles/admin/web/lar/RolesAdminPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.rolesadmin.lar;
+package com.liferay.roles.admin.web.lar;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
