+++ modules/apps/roles-admin/roles-admin-web/src/com/liferay/roles/admin/web/portlet/RolesAdminPortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.rolesadmin;
+package com.liferay.roles.admin.web.portlet;
 
 import com.liferay.portal.DuplicateRoleException;
 import com.liferay.portal.NoSuchRoleException;
